[BE] (refactor) Controller 구현
초기 설정된 Controller 기능별 세분화하여 매핑, 컨트롤러 또한 @RequestMapping사용하여 매핑, 현재 저장되있는 아이템 전부 리스트에 넣어 반환하는 기능 itemList구현,
테스트용 초기 데이터 설정을 위한 @PostConstruct 사용
