--- 4 WEEK/item-service/item-service/src/main/java/hello/itemservice/controller/Controller.java
@@ -1,16 +1,0 @@
-package hello.itemservice.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@org.springframework.stereotype.Controller
-public class Controller {
-    @GetMapping("/itemList")
-    public String itemList() {
-        return "itemList";
-    }
-    @GetMapping("/itemAddForm")
-    public String itemAdd() {
-        return "itemAddForm";
-    }
-}
